+++ src/main/java/remonone/nftilation/game/ingame/services/IPurchasableService.java
@@ -5,6 +5,6 @@ import org.bukkit.entity.Player;
 public interface IPurchasableService {
     
     String getServiceName();
-    void OnPurchase(Player buyer, int price);
+    void OnPurchase(Player buyer, float price);
     
 }
